Rename deepGetter to deepGet
